rtpext: change id unsigned -> uint8_t

+++ src/core.h
@@ -223,7 +223,7 @@ struct rtpext {
 
 
 int rtpext_hdr_encode(struct mbuf *mb, size_t num_bytes);
-int rtpext_encode(struct mbuf *mb, unsigned id, size_t len,
+int rtpext_encode(struct mbuf *mb, uint8_t id, size_t len,
 		  const uint8_t *data);
 int rtpext_decode(struct rtpext *ext, struct mbuf *mb);
